requirements.txt
PySide6==6.7.2
PyYAML==6.0.2
selenium==4.21.0
loguru==0.7.2
pytest==8.3.3
playwright==1.47.0